Refactor: auth에 있는 utils, interfaces common으로 이동
