+++ src/domain/auth/users/users.controller.ts
@@ -5,17 +5,15 @@ import {
   Get,
   Param,
   Patch,
-  Post,
   Req,
   UseGuards,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { JwtAuthGuard } from '@/domain/auth/guards/jwt-auth.guard';
-import RequestWithUser from '@/domain/auth/interfaces/request-with-user.interface';
+import RequestWithUser from '@/common/interfaces/request-with-user.interface';
 import { OneTimeAuthGuard } from '@/domain/auth/guards/one-time-auth.guard';
 import { ChangePasswordDto } from './dto/change-password.dto';
-import { RegisterFcmTokenDto } from '@/domain/auth/users/dto/register-fcm-token.dto';
 import { RedisCacheService } from '@/common';
 import { HttpResponse } from '@/common/dto/http-response';
